requirements.txt
alembic==0.9.5
aniso8601==1.3.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5
coveralls==1.2.0
docopt==0.6.2
ecdsa==0.13
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3.1
py==1.4.34
pycodestyle==2.3.1
pycrypto==2.6.1
pytest==3.2.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
python-jose==2.0.2
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.12.2
